Fix bug if you press multiple operators
If you pressed a second operator without pressing another number the
calculator would apply the operation function anyway to the numbers
stored in A and in displayContent. I put in place flags that stopped
the operate functions being called if an input was required.

Also applied a check on the equal sign to make sure that numbers had
been entered before it carried out its functionality as this could
cause a crash.